Create a basic structure for the project
For linting and testing, pre-commit is used. This is useful because you
can use it directly to check if your code passes the continuous
integration checks even before it is committed.

I also set up a GitHub workflow that performs the pre-commit checks on
the GitHub server, to make sure that no bad code gets pushed anyways.

To manage dependencies, I used 'poetry'. This is used in the GipHouse
website and in Job's original project.

For the Python code, I generated a bit of Django boilerplate using
`django-admin startproject website`, as described in the [official
Django documentation].

I created a .gitignore file that is based on the default GitHub
.gitignore for Python files, but with the irrelevant parts removed.

Finally, I created a README that describes how to get started with
developing.

[official Django documentation]: https://docs.djangoproject.com/en/4.0/intro/tutorial01/