Unhardcode the ledgers
This commit creates a model for ledgers. This way, there is no need to
hardcode them. Its definition is now a bit shorter.

Note that sync_database now creates a Ledger if it encounters a
DocumentLine in which an unknown ledger-id is used. This way, the user
will only have to correctly add the correct ledger-kinds to ledger-ids.

Also, I added a fixture for the ledger ID's as they were previously
hardcoded.